divide files
