+++ test/widget_test.dart
@@ -7,7 +7,6 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
 import 'package:bus_hexa/main.dart';
 
 void main() {
